change directory by domain
