--- src/main/java/com/example/mailService/domain/dto/EmailDto.java
@@ -1,33 +1,0 @@
-package com.example.mailService.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import javax.persistence.Column;
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-public class EmailDto {
-
-    private Long id;
-
-    private String emailFrom;
-
-    private String emailTo;
-
-    private String emailToList;
-
-    private String emailCcList;
-
-    private String emailBccList;
-
-    private String subject;
-
-    private String content;
-
-    private LocalDateTime dateTimeSend;
-
-    private LocalDateTime dateTimeReceive;
-}
