--- src/main/java/com/example/mailService/domain/dto/UserEmailInfoCreateDto.java
@@ -1,38 +1,0 @@
-package com.example.mailService.domain.dto;
-
-import com.example.mailService.domain.entity.User;
-import com.example.mailService.domain.entity.UserEmailInfo;
-import lombok.Builder;
-import lombok.Getter;
-
-import javax.persistence.Column;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-@Getter
-@Builder
-public class UserEmailInfoCreateDto {
-
-    private String email;
-
-    private String username;
-
-    private String password;
-
-    private String smtpHost;
-
-    private Long smtpPort;
-
-    private User user;
-
-    public UserEmailInfo toEntity() {
-        return UserEmailInfo.builder()
-                .email(email)
-                .username(username)
-                .password(password)
-                .smtpHost(smtpHost)
-                .smtpPort(smtpPort)
-                .user(user)
-                .build();
-    }
-}
