--- src/main/java/com/example/mailService/domain/dto/UserLoginDto.java
@@ -1,15 +1,0 @@
-package com.example.mailService.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class UserLoginDto {
-
-    private String username;
-
-    private String password;
-}
