--- src/main/java/com/example/mailService/domain/dto/UserSignUpDto.java
@@ -1,17 +1,0 @@
-package com.example.mailService.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class UserSignUpDto {
-
-    private String username;
-
-    private String email;
-
-    private String password;
-}
