--- src/main/java/com/example/mailService/domain/entity/BaseEntity.java
@@ -1,33 +1,0 @@
-package com.example.mailService.domain.entity;
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public class BaseEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime dateTimeCreated;
-
-    @LastModifiedDate
-    private LocalDateTime dateTimeLastModified;
-
-    @CreatedBy
-    @Column(updatable = false)
-    private Long createdId;
-
-    @LastModifiedBy
-    private Long modifiedId;
-}
