--- src/main/java/com/example/mailService/domain/entity/Email.java
@@ -1,45 +1,0 @@
-package com.example.mailService.domain.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class Email extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Column(nullable = false)
-    private Long userId;
-
-    @Column(nullable = false)
-    private String emailFrom;
-
-    @Column(nullable = false)
-    private String emailTo;
-
-    private String emailToList;
-
-    private String emailCcList;
-
-    private String emailBccList;
-
-    @Column(nullable = false)
-    private String subject;
-
-    private String content;
-
-    private LocalDateTime dateTimeSend;
-
-    private LocalDateTime dateTimeReceive;
-}
