--- src/main/java/com/example/mailService/domain/entity/User.java
@@ -1,32 +1,0 @@
-package com.example.mailService.domain.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class User extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String username;
-
-    @Column(unique = true, nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Enumerated(EnumType.STRING)
-    private UserRole role = UserRole.ROLE_USER;
-}
