--- src/main/java/com/example/mailService/domain/entity/UserRole.java
@@ -1,12 +1,0 @@
-package com.example.mailService.domain.entity;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum UserRole implements GrantedAuthority {
-    ROLE_ADMIN, ROLE_USER;
-
-    @Override
-    public String getAuthority() {
-        return name();
-    }
-}
