+++ src/main/java/com/example/mailService/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.example.mailService.repository;
 
-import com.example.mailService.domain.entity.User;
+import com.example.mailService.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
