+++ src/main/java/com/example/mailService/user/dto/JwtTokenDto.java
@@ -1,0 +1,11 @@
+package com.example.mailService.user.dto;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class JwtTokenDto {
+    private String token;
+}
