+++ src/main/java/com/example/mailService/user/dto/UserLoginDto.java
@@ -1,0 +1,15 @@
+package com.example.mailService.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class UserLoginDto {
+
+    private String username;
+
+    private String password;
+}
