+++ src/main/java/com/example/mailService/user/entity/UserRole.java
@@ -1,0 +1,12 @@
+package com.example.mailService.user.entity;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum UserRole implements GrantedAuthority {
+    ROLE_ADMIN, ROLE_USER;
+
+    @Override
+    public String getAuthority() {
+        return name();
+    }
+}
