+++ src/test/java/com/example/mailService/base/BaseTestSetup.java
@@ -1,7 +1,7 @@
 package com.example.mailService.base;
 
-import com.example.mailService.domain.entity.User;
-import com.example.mailService.domain.entity.UserRole;
+import com.example.mailService.user.entity.User;
+import com.example.mailService.user.entity.UserRole;
 import com.example.mailService.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
